+++ backend/routes/utils/sessionValidators.js
@@ -1,0 +1,12 @@
+const { handleValidationErrors } = require('../../utils/validation');
+
+const validateLogin = [
+  check('credential')
+    .exists({ checkFalsy: true })
+    .notEmpty()
+    .withMessage('Please provide a valid email or username.'),
+  check('password')
+    .exists({ checkFalsy: true })
+    .withMessage('Please provide a password.'),
+  handleValidationErrors,
+];
